requirements.txt
jobspy>=0.29.0
pandas>=2.1.1
python-decouple>=3.8
